feat: /users/meに管理画面ボタンを追加し、/adminを設定一覧ページに変更
- /users/meページのヘッダー右上に「管理画面」ボタンを追加（管理者権限がある場合のみ表示）
- /adminページを設定一覧ページに変更（従来の/admin/settingの機能を統合）
- /admin/settingページを削除
- Header コンポーネントに右側アクション表示機能を追加
- AdminBottomBar、wallet、membersページの戻り先を/adminに変更
- 国際化対応（users.profileHeader.adminButtonを追加）
