+++ src/messages/ja/users.json
@@ -1,5 +1,6 @@
 {
   "users.profileHeader.settingsButton": "設定",
+  "users.profileHeader.adminButton": "管理画面",
   "users.profileHeader.readMore": "もっと見る",
   "users.profileView.certificatesTitle": "証明書",
   "users.portfolio.sectionTitle": "これまでの関わり",
